AI Implemented! Comments on Posts
Make sure to add the API_KEY to the .env, as well as docker compose down and up, whatever is needed.
Will give the response underneath the user input for now, may make it actually make comments as a user or something maybe. But for now this is a start.